move modules to /modules
