+++ public/index.html
@@ -94,10 +94,16 @@
       type="image/png"
       href="/images/favicon.png"
     />
-    <!-- <script async="false" type="text/javascript" src="importmap.js"></script> -->
   </head>
   <body>
-    <!-- <div id="importmap"></div> -->
+    <script type="importmap">
+      {
+        "imports": {
+          "@popperjs/core": "https://cdn.jsdelivr.net/npm/@popperjs/core@2.11.6/dist/umd/popper.min.js",
+          "bootstrap.esm.min.js": "https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.esm.min.js"
+        }
+      }
+    </script>
     <!--[if lt IE 7]>
       <p>
         You are using an <strong>outdated</strong> browser. Please
